Update with changes on trunk to r13235
git-svn-id: file:///home/xapian-svn/xapian/branches/geospatial@13236 6b2f1b92-a4f3-0310-943f-de3cbe41748b

+++ xapian-bindings/python/pythontest2.py
@@ -939,7 +939,7 @@ def test_postingsource():
             self.current = -1
 
         def get_termfreq_min(self): return 0
-        def get_termfreq_est(self): return self.max / 2
+        def get_termfreq_est(self): return int(self.max / 2)
         def get_termfreq_max(self): return self.max
         def next(self, minweight):
             self.current += 2
